+++ packages/@tailwindcss-node/src/compile.ts
@@ -37,7 +37,7 @@ export async function __unstable__loadDesignSystem(css: string, { base }: { base
   })
 }
 
-async function loadModule(id: string, base: string, onDependency: (path: string) => void) {
+export async function loadModule(id: string, base: string, onDependency: (path: string) => void) {
   if (id[0] !== '.') {
     let resolvedPath = await resolveJsId(id, base)
     if (!resolvedPath) {
@@ -73,6 +73,8 @@ async function loadModule(id: string, base: string, onDependency: (path: string)
 async function loadStylesheet(id: string, base: string, onDependency: (path: string) => void) {
   let resolvedPath = await resolveCssId(id, base)
   if (!resolvedPath) throw new Error(`Could not resolve '${id}' from '${base}'`)
+
+  onDependency(resolvedPath)
 
   if (typeof globalThis.__tw_readFile === 'function') {
     let file = await globalThis.__tw_readFile(resolvedPath, 'utf-8')