+++ packages/@tailwindcss-upgrade/src/index.test.ts
@@ -15,6 +15,7 @@ it('should print the input as-is', async () => {
           /* below */
         }
       `,
+      {},
       expect.getState().testPath,
     ),
   ).toMatchInlineSnapshot(`
@@ -29,41 +30,44 @@ it('should print the input as-is', async () => {
 
 it('should migrate a stylesheet', async () => {
   expect(
-    await migrateContents(css`
-      @tailwind base;
+    await migrateContents(
+      css`
+        @tailwind base;
 
-      html {
-        overflow: hidden;
-      }
-
-      @tailwind components;
-
-      .a {
-        z-index: 1;
-      }
-
-      @layer components {
-        .b {
-          z-index: 2;
+        html {
+          overflow: hidden;
         }
-      }
 
-      .c {
-        z-index: 3;
-      }
+        @tailwind components;
 
-      @tailwind utilities;
-
-      .d {
-        z-index: 4;
-      }
-
-      @layer utilities {
-        .e {
-          z-index: 5;
+        .a {
+          z-index: 1;
         }
-      }
-    `),
+
+        @layer components {
+          .b {
+            z-index: 2;
+          }
+        }
+
+        .c {
+          z-index: 3;
+        }
+
+        @tailwind utilities;
+
+        .d {
+          z-index: 4;
+        }
+
+        @layer utilities {
+          .e {
+            z-index: 5;
+          }
+        }
+      `,
+      {},
+    ),
   ).toMatchInlineSnapshot(`
     "@import 'tailwindcss';
 
@@ -103,18 +107,50 @@ it('should migrate a stylesheet', async () => {
 
 it('should migrate a stylesheet (with imports)', async () => {
   expect(
-    await migrateContents(css`
-      @import 'tailwindcss/base';
-      @import './my-base.css';
-      @import 'tailwindcss/components';
-      @import './my-components.css';
-      @import 'tailwindcss/utilities';
-      @import './my-utilities.css';
-    `),
+    await migrateContents(
+      css`
+        @import 'tailwindcss/base';
+        @import './my-base.css';
+        @import 'tailwindcss/components';
+        @import './my-components.css';
+        @import 'tailwindcss/utilities';
+        @import './my-utilities.css';
+      `,
+      {},
+    ),
   ).toMatchInlineSnapshot(`
     "@import 'tailwindcss';
     @import './my-base.css' layer(base);
     @import './my-components.css' layer(components);
     @import './my-utilities.css' layer(utilities);"
+  `)
+})
+
+it('should migrate a stylesheet (with preceding rules that should be wrapped in an `@layer`)', async () => {
+  expect(
+    await migrateContents(
+      css`
+        @charset "UTF-8";
+        @layer foo, bar, baz;
+        /**! My license comment */
+        html {
+          color: red;
+        }
+        @tailwind base;
+        @tailwind components;
+        @tailwind utilities;
+      `,
+      {},
+    ),
+  ).toMatchInlineSnapshot(`
+    "@charset "UTF-8";
+    @layer foo, bar, baz;
+    /**! My license comment */
+    @import 'tailwindcss';
+    @layer base {
+      html {
+        color: red;
+      }
+    }"
   `)
 })