+++ packages/tailwindcss/src/compat/config/types.ts
@@ -79,6 +79,15 @@ export interface ResolvedConfig {
   prefix: string
 }
 
+// `blocklist` support
+export interface UserConfig {
+  blocklist?: string[]
+}
+
+export interface ResolvedConfig {
+  blocklist: string[]
+}
+
 // `important` support
 export interface UserConfig {
   important?: boolean | string